refactor(ui): make app_component a directory module, move AppState out
app_component.rs is 1300 lines with three functions accounting for two
thirds of it. Turning it into a directory gives those a place to land,
starting with the easiest one: AppState is a plain data holder with no
dependency on AppComponent at all.

AppState stays reachable at ui::app_component::AppState through a re-export,
so nothing outside had to change. The tests split the same way, mirroring
the source tree like tests/ui/core and tests/ui/components already do.

Pure move, no behaviour change. handle_global_key and handle_app_action
follow separately.
