+++ src/ui/app_component/state.rs
@@ -1,0 +1,34 @@
+//! The data the app renders, kept apart from the components that draw it.
+
+use crate::entities::{label, project, section, task};
+use crate::ui::core::SidebarSelection;
+
+/// Application state separate from UI concerns
+#[derive(Debug, Clone, Default)]
+pub struct AppState {
+    pub projects: Vec<project::Model>,
+    pub tasks: Vec<task::Model>,
+    pub labels: Vec<label::Model>,
+    pub sections: Vec<section::Model>,
+    pub sidebar_selection: SidebarSelection,
+    pub loading: bool,
+    pub show_help: bool,
+    /// didnt we just got rid of custom scrolling ?
+    pub help_scroll_offset: usize,
+}
+
+impl AppState {
+    /// Update all data at once
+    pub fn update_data(
+        &mut self,
+        projects: Vec<project::Model>,
+        labels: Vec<label::Model>,
+        sections: Vec<section::Model>,
+        tasks: Vec<task::Model>,
+    ) {
+        self.projects = projects;
+        self.labels = labels;
+        self.sections = sections;
+        self.tasks = tasks;
+    }
+}
